+++ src/plugins/event_bus/public/event_bus.dispatch.test.ts
@@ -1,0 +1,45 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the "Elastic License
+ * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
+ * Public License v 1"; you may not use this file except in compliance with, at
+ * your election, the "Elastic License 2.0", the "GNU Affero General Public
+ * License v3.0 only", or the "Server Side Public License, v 1".
+ */
+
+import type { Subscription } from 'rxjs';
+
+import { mockInitialState, mockTestSlice } from './__mocks__/event_bus_mocks';
+import { EventBus } from './event_bus';
+
+describe('EventBus', () => {
+  let eventBus: EventBus<typeof mockTestSlice>;
+  let mockSubscriber: jest.Mock;
+  let subscription: Subscription;
+
+  beforeEach(() => {
+    eventBus = new EventBus(mockTestSlice);
+    mockSubscriber = jest.fn();
+    subscription = eventBus.subscribe(mockSubscriber);
+  });
+
+  afterEach(() => {
+    subscription.unsubscribe();
+    eventBus.dispose();
+  });
+
+  describe('dispatch', () => {
+    it('should work with the dispatch method', () => {
+      // Reset mock to clear initial call
+      mockSubscriber.mockReset();
+
+      eventBus.dispatch(mockTestSlice.actions.increment());
+
+      expect(mockSubscriber).toHaveBeenCalledTimes(1);
+      expect(mockSubscriber).toHaveBeenLastCalledWith({
+        ...mockInitialState(),
+        counter: 1,
+      });
+    });
+  });
+});
