--- src/plugins/event_bus/public/event_bus.test.ts
@@ -1,471 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the "Elastic License
- * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
- * Public License v 1"; you may not use this file except in compliance with, at
- * your election, the "Elastic License 2.0", the "GNU Affero General Public
- * License v3.0 only", or the "Server Side Public License, v 1".
- */
-
-import { configureStore, createAsyncThunk, createSlice, PayloadAction } from '@reduxjs/toolkit';
-import type { Subscription } from 'rxjs';
-
-import { mockInitialState, mockTestSlice, type MockTestState } from './__mocks__/event_bus_mocks';
-import { EventBus } from './event_bus';
-
-describe('EventBus', () => {
-  let eventBus: EventBus<typeof mockTestSlice>;
-  let mockSubscriber: jest.Mock;
-  let subscription: Subscription;
-
-  beforeEach(() => {
-    eventBus = new EventBus(mockTestSlice);
-    mockSubscriber = jest.fn();
-    subscription = eventBus.subscribe(mockSubscriber);
-  });
-
-  afterEach(() => {
-    subscription.unsubscribe();
-    eventBus.dispose();
-  });
-
-  describe('initialization', () => {
-    it('should initialize with the correct initial state', () => {
-      // subscribing (via beforeEach) should trigger an initial state update
-      expect(mockSubscriber).toHaveBeenCalledTimes(1);
-      expect(mockSubscriber).toHaveBeenLastCalledWith(mockInitialState());
-    });
-
-    it('should wrap all slice actions', () => {
-      expect(Object.keys(eventBus.actions)).toEqual(Object.keys(mockTestSlice.actions));
-    });
-  });
-
-  describe('state updates', () => {
-    it('should update state when actions are dispatched', () => {
-      // Reset mock to clear initial call
-      mockSubscriber.mockReset();
-
-      eventBus.actions.increment();
-      eventBus.actions.setText('test');
-
-      expect(mockSubscriber).toHaveBeenCalledTimes(2);
-      expect(mockSubscriber).toHaveBeenLastCalledWith({
-        ...mockInitialState(),
-        counter: 1,
-        text: 'test',
-      });
-    });
-
-    it('should handle complex state updates', () => {
-      // Reset mock to clear initial call
-      mockSubscriber.mockReset();
-
-      eventBus.actions.addItem('item1');
-      eventBus.actions.addItem('item2');
-      eventBus.actions.toggleFlag();
-
-      expect(mockSubscriber).toHaveBeenCalledTimes(3);
-      expect(mockSubscriber).toHaveBeenLastCalledWith({
-        ...mockInitialState(),
-        flag: true,
-        data: {
-          items: ['item1', 'item2'],
-        },
-      });
-    });
-
-    it('should emit state in correct sequence', () => {
-      // Reset mock to clear initial call
-      mockSubscriber.mockReset();
-
-      eventBus.actions.increment();
-      eventBus.actions.setText('test');
-      eventBus.actions.reset();
-
-      expect(mockSubscriber).toHaveBeenCalledTimes(3);
-      expect(mockSubscriber.mock.calls[0][0]).toEqual({ ...mockInitialState(), counter: 1 });
-      expect(mockSubscriber.mock.calls[1][0]).toEqual({
-        ...mockInitialState(),
-        counter: 1,
-        text: 'test',
-      });
-      expect(mockSubscriber.mock.calls[2][0]).toEqual(mockInitialState());
-    });
-
-    it('should reset state correctly', () => {
-      // Reset mock to clear initial call
-      mockSubscriber.mockReset();
-
-      eventBus.actions.increment();
-      eventBus.actions.setText('test');
-      eventBus.actions.reset();
-
-      expect(mockSubscriber).toHaveBeenCalledTimes(3);
-      expect(mockSubscriber).toHaveBeenLastCalledWith(mockInitialState());
-    });
-  });
-
-  describe('selectors', () => {
-    let selectorSubscription: Subscription;
-
-    afterEach(() => {
-      if (selectorSubscription) {
-        selectorSubscription.unsubscribe();
-      }
-    });
-
-    it('should work with selectors', () => {
-      const counterSelector = (state: MockTestState) => state.counter;
-      const mockCounterSubscriber = jest.fn();
-
-      selectorSubscription = eventBus.subscribe(mockCounterSubscriber, counterSelector);
-
-      eventBus.actions.increment();
-
-      expect(mockCounterSubscriber).toHaveBeenCalledTimes(2); // Initial + increment
-      expect(mockCounterSubscriber.mock.calls[0][0]).toBe(0);
-      expect(mockCounterSubscriber.mock.calls[1][0]).toBe(1);
-    });
-
-    it("should not emit when selected value hasn't changed", () => {
-      const textSelector = (state: MockTestState) => state.text;
-      const mockTextSubscriber = jest.fn();
-
-      selectorSubscription = eventBus.subscribe(mockTextSubscriber, textSelector);
-
-      eventBus.actions.setText('test');
-      eventBus.actions.setText('test'); // Same value
-      eventBus.actions.increment(); // Different property
-
-      expect(mockTextSubscriber).toHaveBeenCalledTimes(2); // Initial + first setText
-      expect(mockTextSubscriber.mock.calls[0][0]).toBe('');
-      expect(mockTextSubscriber.mock.calls[1][0]).toBe('test');
-    });
-
-    it('should handle multiple actions correctly', () => {
-      const itemsSelectors = (state: MockTestState) => state.data.items;
-      const mockItemsSubscriber = jest.fn();
-
-      eventBus.subscribe(mockItemsSubscriber, itemsSelectors);
-
-      expect(mockItemsSubscriber).toHaveBeenCalledTimes(1);
-      expect(mockItemsSubscriber).toHaveBeenLastCalledWith([]);
-
-      eventBus.actions.addItem('a');
-      eventBus.actions.addItem('b');
-
-      expect(mockItemsSubscriber).toHaveBeenCalledTimes(3);
-      expect(mockItemsSubscriber).toHaveBeenLastCalledWith(['a', 'b']);
-    });
-  });
-
-  describe('custom equality checks', () => {
-    it('should handle NaN values correctly', () => {
-      const nanSlice = createSlice({
-        name: 'nan',
-        initialState: { value: 0 },
-        reducers: {
-          setNaN: (state) => {
-            state.value = NaN;
-          },
-        },
-      });
-
-      const nanEventBus = new EventBus(nanSlice);
-      const mockNaNSubscriber = jest.fn();
-
-      const nanSubscription = nanEventBus.subscribe(mockNaNSubscriber);
-
-      // this should cause a state update
-      nanEventBus.actions.setNaN();
-      // this should not cause an update since the state is the same
-      nanEventBus.actions.setNaN();
-
-      expect(mockNaNSubscriber).toHaveBeenCalledTimes(2); // Initial + 1 NaN action
-      expect(mockNaNSubscriber.mock.calls[0][0].value).toBe(0);
-      expect(Number.isNaN(mockNaNSubscriber.mock.calls[1][0].value)).toBe(true);
-
-      nanSubscription.unsubscribe();
-      nanEventBus.dispose();
-    });
-  });
-
-  describe('cleanup', () => {
-    it('should clean up subscriptions on dispose', () => {
-      // Create a new event bus for this test since we'll be disposing it
-      const testEventBus = new EventBus(mockTestSlice);
-      const cleanupMockSubscriber = jest.fn();
-
-      testEventBus.subscribe(cleanupMockSubscriber);
-
-      // Reset mock to clear initial call
-      cleanupMockSubscriber.mockReset();
-
-      testEventBus.dispose();
-      testEventBus.actions.increment();
-
-      expect(cleanupMockSubscriber).not.toHaveBeenCalled();
-    });
-
-    it('should complete the subject on dispose', (done) => {
-      // Create a new event bus for this test since we'll be disposing it
-      const testEventBus = new EventBus(mockTestSlice);
-
-      testEventBus.subject.subscribe({
-        complete: () => {
-          done();
-        },
-      });
-
-      testEventBus.dispose();
-    });
-  });
-
-  describe('error handling', () => {
-    it('should propagate errors from reducers through subscription', (done) => {
-      const testError = new Error('Subscription error test');
-      const errorSlice = createSlice({
-        name: 'error',
-        initialState: { value: 0 },
-        reducers: {
-          throwError: () => {
-            throw testError;
-          },
-        },
-      });
-
-      const errorEventBus = new EventBus(errorSlice);
-      const mockCallback = jest.fn();
-      const mockSelector = jest.fn();
-      const mockErrorHandler = jest.fn();
-
-      // Subscribe with an error handler
-      const errorSubscription = errorEventBus.subscribe(
-        mockCallback,
-        mockSelector,
-        mockErrorHandler
-      );
-
-      try {
-        errorEventBus.actions.throwError();
-      } catch (e) {
-        // In the real implementation, errors in reducers should be caught
-        // and propagated through the observable error channel
-        // If this catch block executes, the test should fail because
-        // the error wasn't properly propagated
-        errorSubscription.unsubscribe();
-        errorEventBus.dispose();
-        done.fail('Error was thrown instead of being propagated through the observable');
-      }
-
-      expect(mockErrorHandler).toHaveBeenCalledWith(testError);
-      done();
-    });
-
-    it('should have error handling support in subscribe method', () => {
-      // This test verifies that the subscribe method accepts error handlers
-      const mockErrorHandler = jest.fn();
-
-      // Create a modified EventBus implementation that tests error handling
-      class TestErrorEventBus extends EventBus<typeof mockTestSlice> {
-        triggerError(error: Error) {
-          this.subject.error(error);
-        }
-      }
-
-      const errorBus = new TestErrorEventBus(mockTestSlice);
-
-      const errorSubscription = errorBus.subscribe(
-        () => {}, // next handler
-        undefined, // no selector
-        mockErrorHandler // error handler
-      );
-
-      const testError = new Error('Subscription error test');
-      errorBus.triggerError(testError);
-
-      expect(mockErrorHandler).toHaveBeenCalledWith(testError);
-
-      errorSubscription.unsubscribe();
-      errorBus.dispose();
-    });
-  });
-
-  describe('subscription management', () => {
-    it('should handle multiple subscriptions correctly', () => {
-      // We already have one subscription set up in beforeEach
-      const mockSubscriber2 = jest.fn();
-
-      const subscription2 = eventBus.subscribe(mockSubscriber2);
-
-      // Reset the mocks to clear initial calls
-      mockSubscriber.mockReset();
-      mockSubscriber2.mockReset();
-
-      eventBus.actions.increment();
-
-      expect(mockSubscriber).toHaveBeenCalledTimes(1); // increment
-      expect(mockSubscriber2).toHaveBeenCalledTimes(1); // increment
-
-      // Unsubscribe second subscription
-      subscription2.unsubscribe();
-
-      eventBus.actions.setText('test');
-
-      // First subscriber continues to receive updates
-      expect(mockSubscriber).toHaveBeenCalledTimes(2);
-      // Second subscriber should not be called anymore
-      expect(mockSubscriber2).toHaveBeenCalledTimes(1);
-    });
-
-    it('should unsubscribe properly', () => {
-      // Create a new subscription for this test
-      const testMockSubscriber = jest.fn();
-      const testSubscription = eventBus.subscribe(testMockSubscriber);
-
-      // Reset mock to clear initial call
-      testMockSubscriber.mockReset();
-
-      testSubscription.unsubscribe();
-
-      eventBus.actions.increment();
-
-      expect(testMockSubscriber).not.toHaveBeenCalled();
-
-      // Our main subscription from beforeEach should still work
-      expect(mockSubscriber).toHaveBeenCalled();
-    });
-  });
-
-  describe('dispatch', () => {
-    it('should work with the dispatch method', () => {
-      // Reset mock to clear initial call
-      mockSubscriber.mockReset();
-
-      eventBus.dispatch(mockTestSlice.actions.increment());
-
-      expect(mockSubscriber).toHaveBeenCalledTimes(1);
-      expect(mockSubscriber).toHaveBeenLastCalledWith({
-        ...mockInitialState(),
-        counter: 1,
-      });
-    });
-  });
-
-  describe('async thunks', () => {
-    it('should handle async thunks', async () => {
-      jest.useFakeTimers();
-
-      // Create an async action that updates the state after a delay
-      const asyncAction = createAsyncThunk('test/asyncAction', async (_, thunkApi) => {
-        setTimeout(() => {
-          thunkApi.dispatch(asyncSlice.actions.setData('the-async-data'));
-        }, 100);
-      });
-
-      // State initializer
-      const getInitialState: () => { data: null | string; isLoading: boolean } = () => ({
-        data: null,
-        isLoading: false,
-      });
-
-      // Create a slice with an async action
-      const asyncSlice = createSlice({
-        name: 'async',
-        initialState: getInitialState(),
-        reducers: {
-          setData: (state, action: PayloadAction<string>) => {
-            state.data = action.payload;
-          },
-        },
-        extraReducers: (builder) => {
-          builder.addCase(asyncAction.pending, (state) => {
-            state.isLoading = true;
-          });
-          builder.addCase(asyncAction.fulfilled, (state) => {
-            state.isLoading = false;
-          });
-        },
-      });
-
-      // Create an event bus for the async slice
-      const asyncEventBus = new EventBus(asyncSlice);
-
-      // Subscribe to the async event bus
-      const asyncMockSubscriber = jest.fn();
-      const asyncSubscription = asyncEventBus.subscribe(asyncMockSubscriber);
-
-      // Initial state should be emitted
-      expect(asyncMockSubscriber).toHaveBeenCalledTimes(1);
-      expect(asyncMockSubscriber).toHaveBeenLastCalledWith(getInitialState());
-
-      // Dispatch the async action
-      const resp = asyncEventBus.dispatch(asyncAction());
-
-      // The state should update to show that the async action is loading
-      expect(asyncMockSubscriber).toHaveBeenCalledTimes(2);
-      expect(asyncMockSubscriber).toHaveBeenLastCalledWith({
-        ...getInitialState(),
-        isLoading: true,
-      });
-
-      // Fast-forward time to allow the async action to complete
-      jest.advanceTimersByTime(500);
-
-      // IMPORTANT: We need to await the thunk's promise to catch the fulfilled state
-      await resp;
-
-      // The state should update to show that the async action has completed
-      expect(asyncMockSubscriber).toHaveBeenCalledTimes(4);
-      expect(asyncMockSubscriber).toHaveBeenLastCalledWith({
-        ...getInitialState(),
-        data: 'the-async-data',
-        isLoading: false,
-      });
-
-      // cleanup
-      asyncSubscription.unsubscribe();
-      asyncEventBus.dispose();
-      jest.useRealTimers();
-    });
-  });
-
-  describe('performance', () => {
-    it('should handle 1000 actions in reasonable time', () => {
-      const start = performance.now();
-
-      for (let i = 0; i < 1000; i++) {
-        eventBus.actions.increment();
-      }
-
-      const duration = performance.now() - start;
-
-      // Set a generous upper bound to avoid CI failures.
-      // 500ms is very generous, actual time is ~15ms but
-      // this should let us track severe regressions.
-      expect(duration).toBeLessThan(500);
-    });
-
-    it('should not be significantly slower than plain Redux', () => {
-      const store = configureStore({ reducer: mockTestSlice.reducer });
-
-      const reduxStart = performance.now();
-      for (let i = 0; i < 1000; i++) {
-        store.dispatch(mockTestSlice.actions.increment());
-      }
-      const reduxDuration = performance.now() - reduxStart;
-
-      const eventBusStart = performance.now();
-      for (let i = 0; i < 1000; i++) {
-        eventBus.actions.increment();
-      }
-      const eventBusDuration = performance.now() - eventBusStart;
-
-      // Assert event bus shouldn't be dramatically slower
-      // 1.5x is a generous factor to account for CI variability
-      // Local baseline is that both eventBus and plain Redux are ~15ms.
-      expect(eventBusDuration).toBeLessThan(reduxDuration * 1.5);
-    });
-  });
-});
